+++ packages/Rfmalloc/src/fmalloc.cpp
@@ -9,5 +9,6 @@
 #include "fmalloc_tensor.inc"
 #include "fmalloc_alp.inc"
 #include "fmalloc_sparse.inc"
+#include "fmalloc_bed.inc"
 #include "fmalloc_insitu.inc"
 #include "fmalloc_api.inc"